+++ src/main/java/com/danielagapov/spawn/Services/UserSearch/IUserSearchService.java
@@ -1,0 +1,16 @@
+package com.danielagapov.spawn.Services.UserSearch;
+
+import com.danielagapov.spawn.DTOs.User.BaseUserDTO;
+import com.danielagapov.spawn.DTOs.User.FriendUser.RecommendedFriendUserDTO;
+import com.danielagapov.spawn.Util.SearchedUserResult;
+
+import java.util.List;
+import java.util.UUID;
+
+public interface IUserSearchService {
+    SearchedUserResult getRecommendedFriendsBySearch(UUID requestingUserId, String searchQuery);
+
+    List<RecommendedFriendUserDTO> getLimitedRecommendedFriendsForUserId(UUID userId);
+
+    List<BaseUserDTO> searchByQuery(String searchQuery);
+}
